requirements.txt
aiofiles==0.8.0
albucore==0.0.13
albumentations==1.4.10
allure-pytest==2.13.5
allure-python-commons==2.13.5
annotated-types==0.7.0
anyio==4.4.0
asgiref==3.5.1
astor==0.8.1
asttokens==2.4.1
async-timeout==4.0.3
atomicwrites==1.4.0
attrs==24.2.0
backcall==0.2.0
bcrypt==4.2.0
beautifulsoup4==4.12.3
bleach==6.1.0
blinker==1.4
Brotli==1.0.9
cachetools==5.5.0
certifi==2021.10.8
cffi==1.15.0
chardet==4.0.0
charset-normalizer==2.0.7
click==8.1.3
colorama==0.4.4
colorlog==6.6.0
contourpy==1.3.0
crypto==1.4.1
cryptography==36.0.0
cssselect==1.2.0
cssutils==2.11.1
cycler==0.12.1
Cython==3.0.11
decorator==5.1.1
defusedxml==0.7.1
DingtalkChatbot==1.5.3
docopt==0.6.2
et-xmlfile==1.1.0
eval_type_backport==0.2.0
exceptiongroup==1.2.2
execnet==1.9.0
executing==2.1.0
exif==1.6.0
Faker==9.8.3
fastjsonschema==2.20.0
fire==0.6.0
Flask==2.0.3
fonttools==4.53.1
greenlet==3.1.1
h11==0.13.0
h2==4.1.0
hpack==4.0.0
httpcore==1.0.5
httptools==0.4.0
httpx==0.27.2
hyperframe==6.0.1
idna==3.3
imageio==2.35.1
imgaug==0.4.0
importlib_metadata==8.4.0
importlib_resources==6.4.4
imutils==0.5.4
iniconfig==1.1.0
ipython==8.12.3
itchat==1.3.10
itsdangerous==2.1.2
jedi==0.19.1
Jinja2==3.1.2
joblib==1.4.2
jsonpath==0.82
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
jupyter_client==8.6.2
jupyter_core==5.7.2
jupyterlab_pygments==0.3.0
kaitaistruct==0.9
kiwisolver==1.4.5
lazy_loader==0.4
ldap3==2.9.1
lmdb==1.5.1
loguru==0.7.2
lxml==5.3.0
MarkupSafe==2.1.1
matplotlib==3.9.2
matplotlib-inline==0.1.7
mistune==3.0.2
mitmproxy==8.1.1
more-itertools==10.5.0
msgpack==1.0.3
multidict==6.0.2
Naked==0.1.32
nbclient==0.10.0
nbconvert==7.16.4
nbformat==5.10.4
networkx==3.2.1
numpy==1.26.4
opencv-contrib-python==4.10.0.84
opencv-python==4.10.0.84
opencv-python-headless==4.10.0.84
openpyxl==3.0.9
opt-einsum==3.3.0
packaging==21.3
paddleocr==2.9.1
paddlepaddle==2.6.1
pandocfilters==1.5.1
paramiko==3.5.0
parso==0.8.4
passlib==1.7.4
pathvalidate==3.2.1
pickleshare==0.7.5
pillow==10.4.0
pipreqs==0.5.0
platformdirs==4.2.2
playwright==1.48.0
pluggy==1.5.0
plum-py==0.8.7
premailer==3.10.0
prompt_toolkit==3.0.47
protobuf==3.19.4
psutil==6.1.0
publicsuffix2==2.20191221
pure_eval==0.2.3
py==1.11.0
pyasn1==0.4.8
pyclipper==1.3.0.post5
pycparser==2.21
pydantic==2.10.0
pydantic_core==2.27.0
pyDes==2.0.1
pydivert==2.1.0
pyee==12.0.0
Pygments==2.18.0
PyMySQL==1.0.2
PyNaCl==1.5.0
pyOpenSSL==21.0.0
pyparsing==3.0.6
pyperclip==1.8.2
pypng==0.0.21
PyQRCode==1.2.1
pytest==8.3.3
pytest-base-url==2.1.0
pytest-dependency==0.6.0
pytest-forked==1.3.0
pytest-playwright==0.5.1
pytest-rerunfailures==14.0
pytest-xdist==2.4.0
python-dateutil==2.8.2
python-docx==1.1.2
python-slugify==8.0.4
pywin32==304
PyYAML==5.4.1
pyzmq==26.2.0
rapidfuzz==3.9.7
redis==4.3.6
referencing==0.35.1
requests==2.32.3
requests-toolbelt==0.9.1
rpds-py==0.20.0
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.6
sanic==22.3.1
sanic-routing==22.3.0
scikit-image==0.24.0
scikit-learn==1.5.2
scipy==1.13.1
shapely==2.0.6
shellescape==3.8.1
six==1.16.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.6
sshtunnel==0.4.0
stack-data==0.6.3
termcolor==2.4.0
text-unidecode==1.3
threadpoolctl==3.5.0
tifffile==2024.8.30
tinycss2==1.3.0
toml==0.10.2
tomli==2.0.1
tornado==6.4.1
tqdm==4.66.5
traitlets==5.14.3
typing_extensions==4.12.2
urllib3==1.26.7
urwid==2.1.2
wcwidth==0.2.13
webencodings==0.5.1
websockets==10.3
Werkzeug==2.1.2
win32-setctime==1.1.0
wsproto==1.1.0
xlrd==2.0.1
xlutils==2.0.0
xlwings==0.27.7
xlwt==1.3.0
yagmail==0.15.293
yarg==0.1.9
zipp==3.20.1
zstandard==0.17.0

